+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [`MultiTimeOnlyRange<T>`](/src/Candoumbe.MiscUtilities/Types/MultiTimeOnlyTimeRange.cs) type
 - Added [`MultiDateOnlyRange<T>`](/src/Candoumbe.MiscUtilities/Types/MultiDateOnlyTimeRange.cs) type
 
+## [0.8.1] / 2022-07-16
+- Fixed validating a email pattern using `Like` extension method. 
+
 ## [0.8.0] / 2022-01-26
 - Added `SortBy` extension method for `IEnumerable<T>` type
 - Added `IShuffler` interface and `FisherYatesShuffler` implementation
@@ -72,7 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `ToSnakeCase` to take into account characters like `-`
 - Made [ReplaceVisitor](./src/Candoumbe.MiscUtilities/ReplaceVisitor.cs) public
 
-[Unreleased]: https://github.com/candoumbe/MiscUtilities/compare/0.8.0...HEAD
+[Unreleased]: https://github.com/candoumbe/MiscUtilities/compare/0.8.1...HEAD
+[0.8.1]: https://github.com/candoumbe/MiscUtilities/compare/0.8.0...0.8.1
 [0.8.0]: https://github.com/candoumbe/MiscUtilities/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/candoumbe/MiscUtilities/compare/0.6.3...0.7.0
 [0.6.3]: https://github.com/candoumbe/MiscUtilities/compare/0.6.2...0.6.3